backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-dotenv
google-generativeai
python-multipart
pyserial